finish

+++ www/models.py
@@ -4,7 +4,7 @@
 '''
 Models for user, blog, comment.
 '''
-from www.orm import Model, StringField, BooleanField, TextField, FloatField
+from www.orm import Model, StringField, BooleanField, FloatField, TextField
 
 __author__ = 'Michael Liao'
 
@@ -23,7 +23,7 @@ class User(Model):
     admin = BooleanField()
     name = StringField(ddl='varchar(50)')
     image = StringField(ddl='varchar(500)')
-    created_at = default=time.time
+    created_at = FloatField(default=time.time)
 
 class Blog(Model):
     __table__ = 'blogs'
